refactor : 공통된 로직 모듈화
