+++ auth/signin.html
@@ -58,6 +58,7 @@
         </div>
       </div>
     </div>
-    <script src="auth.js"></script>
+    <script type="module" src="main.js"></script>
+    <script src="./utill/toggle.js"></script>
   </body>
 </html>
