+++ auth/signup.html
@@ -74,6 +74,7 @@
         </div>
       </div>
     </div>
-    <script src="auth.js"></script>
+    <script type="module" src="main.js"></script>
+    <script src="./utill/toggle.js"></script>
   </body>
 </html>
