+++ auth/utill/constant.js
@@ -1,0 +1,16 @@
+export const ERROR_MESSAGE = {
+  signin: {
+    EMPTY_EMAIL: "이메일을 입력해주세요.",
+    EMPTY_PASSWORD: "비밀번호를 입력해주세요.",
+    INVALID_EMAIL: "이메일을 확인해주세요",
+    INVALID_PASSWORD: "비밀번호를 확인해주세요.",
+  },
+  signup: {
+    EMPTY_EMAIL: "이메일을 입력해주세요.",
+    EMPTY_PASSWORD: "비밀번호를 입력해주세요.",
+    INVALID_EMAIL: "올바른 이메일 주소가 아닙니다.",
+    INVALID_PASSWORD: "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요.",
+    EXIST_EMAIL: "이미 존재하는 이메일입니다.",
+    PASSWORD_EQUAL: "비밀번호가 일치하지 않아요.",
+  },
+};
